[#2480] Redesign talkform to resemble quick-reply
This doesn't remove any of the features I targeted in #2480, because we haven't
met the preconditions yet. It just moves things around and adds an icon preview.

- The "From" radio buttons go above the icon preview, taking over from the
  quick-reply's username indicator.
- "More options" is omitted, because we're already here.
- The subjecticon button stays for now, and goes between the subject and the
  quote button.
- The spellcheck, admin post, preformat, and unscreen checkboxes go under the
  message body and before the submit buttons. So do the captcha and edit reason
  fields.
- Other than that, it looks like the quick-reply.

This is gated on the `s2foundation` beta. The implementation isn't related to
the other stuff in that beta, but it sorta fits the theme of "replacing
everything in the comment pages."

Also, the Foundation site skins are the only conditions where a redesign is
urgent: the table-based talkform looks outrageously bad there, and re-doing that
CSS from scratch would have been a waste since we don't want this table beast
long-term anyway.