[#2427] [#2570] Use Foundation in S2 pages
This is gated on a beta feature called `s2foundation`.

In site scheme entry/reply/icons, this changes all of the styling.

In journal-styled pages this doesn't change any visible styling, but it DOES
enable the use of Foundation-based components in those pages. It uses a new
minimal CSS file to set up Foundation's prerequisites without changing anything
important. (Foundation checks for metadata hidden in weird dummy style rules
before it'll let its JavaScript do anything, and also it relies on Modernizr
classes; ironically old-fashioned, but never mind that.)

The beta also enables need_res within components called by journal-styled S2
pages... but only for JavaScript includes, not CSS includes. It's an issue with
S2 eagerly printing stuff, alas.

This commit also adds a new read-only variable to S2 Page objects that indicates
whether the control strip will be shown, because that was the lowest-impact way
to deal with the control strip JS.

Known issues in the beta as of this commit:

- A bunch of JavaScript in journal pages is broken, because it's not being added
  to the "foundation" resource group.
- Site skin entry/reply/icons pages look bad because they don't have the
  appropriate CSS yet.