Nerf an OP `#content ul li` rule in _page-layout-hacks.scss
Using an `#id` selector in a default rule makes a mess in components that need
to override it (`!important`, redundant IDs, etc.). But `id` is also a normal
attribute, and `[attribute] li` can be overridden by any subsequent `.component
li` rule.

This commit also removes a bunch of hacks to work around that rule's rowdiness.