Beta create entries page: Improve preview
The preview feature of beta create entries had some jank and cruft, which was
starting to degrade previews:

- It wasn't loading CSS and JS that we expect to have available in journal
content, so things like image shrinking weren't behaving like they do in real
content.
- It was using an outdated template for site skin entry pages that seemed to
predate all the work to move site-skin comments into S2. (On this one, I opted
to just update the template rather than figure which hoops must be jumped
through in order to invoke the siteviews layout from a non-journal context. If
anyone wants to try and sort that one out, it would probably benefit the new
comment preview as well. But for now this was just getting in my way, so I went
for the quick fix. )