DW::Formats - New helper package for working w/ known markup formats
This part actually came in mid-way through this feature, but I'm moving it to
the front because it sort of serves as a map or a thesis statement for the
rest of the changes.

In short, this changeset is introducing "named markup formats" as a central
interface for dealing with user-entered text.

Each unit of user-provided body text (entry and comment bodies) can include, as
metadata, the name of the markup format it uses. There's already a space
reserved for this: both entries and comments include an `editor` prop,
originally intended for exactly this.

We'll use these named formats for two purposes:

- Picking the correct formatting options for the HTML cleaner. DW stores source
text raw, and then formats it on display; the saved "editor" value will be a
clear signal of exactly what the user expected to have happen when they wrote
the post, so we can freely change our markup behaviors in the present day
(adding `@mentions`, changing Markdown processors, etc.) without worrying about
wreaking havoc on archived content. Formats are versioned, and any significant
change to markup behavior is now expected to result in a new format name.

- Picking the correct format-dependent states and behaviors for the create
entries form and the comment forms. In the initial implementation, this only
affects the selection state of the markup format drop-down. In the future, it
will affect the output of formatting helper buttons for plain-text formats,
and will control whether to display the plain-text or rich text editor. In the
event of a future migration from one rich text editor to another, named formats
will allow us to treat the old one and the new one as separate entities; we
won't need to ensure the new editor can accommodate text created by the old one,
and can even keep both editors operational for some time during the transition.

The initial set of formats is a codification of the formats DW **already uses.**
It's just that they didn't have names before, and were selected by emergent
interactions between the `opt_preformatted` prop, the imported/syndicated
status, Secret Glyphs (`!markdown`), and the editor prop (only supported in
comment-by-email).

Notable behaviors in this new code:

- When building a list of available formats for building a dropdown menu, we
default to only returning the currently "active" list of formats, leaving off
legacy ones that we still need to display right but which we don't want users to
deliberately post with.

- You can provide the current format (for existing content being edited after
the fact) and the user's preferred format. We consider these for choosing the
initial selection state of the menu.

- If the current format is already set to an obsolete/inactive format, we
include that format in the menu, even though we usually wouldn't. This is so you
can edit old posts without having to translate all of the markup behaviors.

- If the user's preference is an obsolete/inactive format, we don't care. They
reset to the default.