LJ::CleanHTML::clean_event/clean_comment - use named formats
- Add a utility function for translating a named markup format to a specific
set of arguments that can be passed to `clean()`. Yeah, it's kind of frustrating
that this isn't in DW::Formats, but keeping the chaos of `clean()` contained to
one file seemed more important.

- Updates clean_event and clean_comment to use named formats from their `editor`
argument if available, and to guess an appropriate format otherwise. The
guessing behavior mostly uses the same heuristics we were already using to
switch off between the cleaner's options, and so it mostly results in the same
behavior as before this patch.

- EXCEPT for two new behaviors: Native site content written before May of 2019
now skips converting `@mentions`, and syndicated content now uses the "extra
raw" version of preformatted HTML. (Though we're not passing in the posting date
yet, so that behavior won't ever fire.)