Pass extra info to clean_event and clean_comment
This updates various callers of clean_event and clean_comment to pass in the
extra info we use to guess the format of old content.

Also, in S2: stop passing a bunch of stuff that LJ::Entry::event_html just
overrides with the same value anyway. (Detritus from a previous code path that
reimplemented event_html by hand.)