NEW USERPROPS: comment_editor and entry_editor2
These userprops are for tracking the user's preferred markup format. I intend
to implement that preference with zero UI -- the last format you used for
posting _new_ (not edited) content is your preference. Don't want that to be
your preference? Just change it next time.

Using a new prop for entry in order to not stomp on the existing prop for the
old RTE.