Talkform and quick-reply: Add formats dropdown (PROD STRING UPDATE)
This adds a formatting dropdown to the reply forms, replacing the old "don't
autoformat" checkbox.

- The dropdown goes between the subject and body fields, and the quote button
moves alongside it to provide some visual balance.

- The dropdown sets a value for the `prop_editor` form field. (Currently unused,
but matching the naming convention of the other property-tied comment form
fields like `prop_opt_preformatted` and `prop_picture_keyword`.)

- The available values and default selected value are deferred to DW::Formats.
To help it decide, we tell it:
    - The _user-preferred_ format is the comment_editor userprop (currently never
    set by anything).
    - The _current_ format is whatever was passed in the `prop_editor` field of
    the submitted form (not passed for comment edits yet, but it'll already
    carry over for "more options", preview, and fixing errors on talkpost_do).

- The help icon is meant to link to a FAQ page all about our formatting options.
On deploy, site owners will need to add the following strings to make that icon
link to the right place:
    - markup.helplink.url
    - markup.helplink.alttext