LJ::Talk::post_comment - Set comment_editor userprop
When a user submits a new (non-edit) comment, we want to update their editor
preference. So I'm doing that at the last mile, in post_comment.

I considered setting it in talkpost_do instead, to skip email and client
comments and only handle POSTs from the reply forms, but unfortunately that
would have left quickreply from reading/recent out in the cold, since it uses
RPC. Open to suggestions on this one; we could just do it in both the Talk
controller and the one RPC call, and leave clients/email alone.