Add `rich1` format for the new RTE
This makes it appear in dropdowns, but it's disabled to start with. We'll use JS
to enable it when it's supported. I haven't done any hardening or smoothing-out
on this flow yet; I'll probably want to modify the html_select helper to allow
totally hidden options or something, and it'll be interesting to figure out what
to do when editing an existing rich1 post on a system that can't use it.